refactor: CLI did
